+++ timepiece/src/main/java/com/appcenter/timepiece/repository/MemberProjectRepository.java
@@ -1,7 +1,7 @@
 package com.appcenter.timepiece.repository;
 
 import com.appcenter.timepiece.domain.MemberProject;
-import com.appcenter.timepiece.repository.impl.MemberProjectRepositoryCustom;
+import com.appcenter.timepiece.repository.customRepository.MemberProjectRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
