+++ timepiece/src/main/java/com/appcenter/timepiece/repository/customRepository/MemberProjectRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.appcenter.timepiece.repository.impl;
+package com.appcenter.timepiece.repository.customRepository;
 
 import com.appcenter.timepiece.domain.MemberProject;
 import org.springframework.data.domain.Page;
