+++ timepiece/src/main/java/com/appcenter/timepiece/repository/impl/MemberProjectRepositoryImpl.java
@@ -1,10 +1,10 @@
-package com.appcenter.timepiece.repository.customRepository;
+package com.appcenter.timepiece.repository.impl;
 
 import com.appcenter.timepiece.domain.MemberProject;
 import com.appcenter.timepiece.generated.com.appcenter.timepiece.domain.QCover;
 import com.appcenter.timepiece.generated.com.appcenter.timepiece.domain.QMemberProject;
 import com.appcenter.timepiece.generated.com.appcenter.timepiece.domain.QProject;
-import com.appcenter.timepiece.repository.impl.MemberProjectRepositoryCustom;
+import com.appcenter.timepiece.repository.customRepository.MemberProjectRepositoryCustom;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
